move the LanguagesDropdownAdapter to the base ui module to reuse it elsewhere
